refactor: break down accounts.ts into focused, maintainable modules
PROBLEM ANALYSIS:
- accounts.ts was a monolithic 2,317-line file with 33 HTTP endpoints
- Multiple types defined inline instead of in proper type directories
- Violated single responsibility principle and hindered maintainability
- Difficult to navigate, test, and maintain in team development

SOLUTION IMPLEMENTED:
1. TYPE ORGANIZATION
   - Created /interfaces/accountInterfaces.ts with all account-related interfaces
   - Extracted 5+ inline types: AccountSearchResult, PublicAccountResponse,
     AccountListResponse, ContactSearchResult, AccountField, etc.
   - Improved type reusability across the application

2. FILE STRUCTURE REFACTORING
   Split into 5 focused modules:
   - accounts-core.ts (585 lines): Core CRUD, search, my-accounts
   - accounts-contacts.ts (334 lines): Contact/user management, roles
   - accounts-resources.ts (485 lines): Teams, leagues, fields, umpires
   - accounts-settings.ts (312 lines): URLs, Twitter, metadata
   - accounts-assets.ts (63 lines): Logo upload/management
   - accounts.ts (29 lines): Main router combining sub-routers

3. CODE QUALITY IMPROVEMENTS
   - Applied DRY principles and removed code duplication
   - Followed single responsibility principle
   - Maintained full backward compatibility (all API endpoints unchanged)
   - Updated test imports to use correct module references
   - Fixed TypeScript types and resolved all linting issues

BENEFITS ACHIEVED:
✅ Improved maintainability - Easy to find and modify specific functionality
✅ Better team development - Reduced merge conflicts, clearer responsibilities
✅ Enhanced testing - Individual functional areas can be tested independently
✅ Cleaner architecture - Follows SOLID principles and DRY patterns
✅ Reusable types - Type definitions shared across application
✅ Easier debugging - Smaller, focused files are easier to navigate

BACKWARD COMPATIBILITY:
- All 33 API endpoints maintain exact same URLs and behavior
- No breaking changes to existing functionality
- Original file backed up as accounts.ts.backup
- All imports and dependencies updated correctly
